+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Current (in progress)
 
-- Nothing yet
+- Topic: add filters in API [#3007](https://github.com/opendatateam/udata/pull/3007)
 
 ## 7.0.6 (2024-03-29)